Move filter GUI registration to GUI package
