+++ mantidimaging/core/utility/registrator.py
